Front end logic

+++ src/main/java/io/nology/employees/employees/EmployeeService.java
@@ -1,7 +1,10 @@
 package io.nology.employees.employees;
 
 import java.time.LocalDate;
+import java.util.Arrays;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 import org.modelmapper.ModelMapper;
@@ -15,6 +18,9 @@ import io.nology.employees.employees.dtos.CreateEmployeeRequest;
 import io.nology.employees.employees.dtos.FindEmployeesQueryDto;
 import io.nology.employees.employees.dtos.UpdateEmployeeRequest;
 import io.nology.employees.employees.entities.Employee;
+import io.nology.employees.employees.entities.EmploymentType;
+import io.nology.employees.employees.entities.Pronouns;
+import io.nology.employees.employees.entities.WorkSetup;
 import io.nology.employees.role.RoleService;
 import io.nology.employees.role.entities.Role;
 
@@ -71,6 +77,9 @@ public class EmployeeService {
         if(result.isEmpty()) {
             return result;
         }
+        if(this.repo.existsByEmailAddressAndIdNot(data.getEmailAddress(), id)) {
+            throw new UnprocessableContentException("This email address is already in use. Please use another");
+        }
         Employee foundEmployee = result.get();
         
         LocalDate newStartDate = data.getStartDate() != null ? data.getStartDate() : foundEmployee.getStartDate();
@@ -100,9 +109,22 @@ public class EmployeeService {
         if(result.isEmpty()) {
             return false;
         }
+        boolean isManager = this.repo.existsByManagerId(id);
+        if(isManager) {
+            throw new UnprocessableContentException("Cannot delete this employee as they are currently a manager of other employee(s)");
+        }
         this.repo.delete(result.get());
         return true;
     }
+
+    public Map<String, List<Map<String, String>>> getAllEnums() {
+        Map<String, List<Map<String, String>>> enums = new HashMap<>();
+        enums.put("pronouns", Arrays.stream(Pronouns.values()).map(e -> Map.of("label", e.getLabel(), "value", e.name())).toList());
+        enums.put("employmentType", Arrays.stream(EmploymentType.values()).map(e -> Map.of("label", e.getLabel(), "value", e.name())).toList());
+        enums.put("workSetup", Arrays.stream(WorkSetup.values()).map(e -> Map.of("label", e.getLabel(), "value", e.name())).toList());
+        return enums;
+    }
+
 
     private void validateDates(LocalDate startDate, LocalDate lastDate) {
         if (startDate != null && lastDate != null && startDate.isAfter(lastDate)) {
@@ -113,15 +135,12 @@ public class EmployeeService {
     private String createUniqueEmail(String firstName, String lastName) {
         String cleanFirst = firstName.toLowerCase().replaceAll("[^a-z0-9]", "");
         String cleanLast = lastName.toLowerCase().replaceAll("[^a-z0-9]", "");
-        
-        String emailBase = cleanFirst + "." + cleanLast;
-        String domain = "@mycompany.com";
 
-        String candidateEmail = emailBase + domain;
+        String candidateEmail = String.format("%s.%s@mycompany.com", cleanFirst, cleanLast);
         int counter = 1;
 
         while(this.repo.existsByEmailAddress(candidateEmail)) {
-            candidateEmail = emailBase + counter + domain;
+            candidateEmail = String.format("%s.%s%s@mycompany.com", cleanFirst, cleanLast, counter);
             counter++;
         }
